SemanticdbDefinition - use correct dialect
