+++ metals-bench/src/main/scala/bench/MetalsBench.scala
@@ -144,7 +144,7 @@ class MetalsBench {
   @Benchmark
   @BenchmarkMode(Array(Mode.SingleShotTime))
   def indexSources(): Unit = {
-    val index = OnDemandSymbolIndex()
+    val index = OnDemandSymbolIndex.empty()
     fullClasspath.entries.foreach(entry =>
       index.addSourceJar(entry, dialects.Scala213)
     )
