+++ tests/cross/src/main/scala/tests/BasePCSuite.scala
@@ -211,7 +211,7 @@ abstract class BasePCSuite extends BaseSuite {
       case NonFatal(e) =>
         println(s"warn: $e")
     }
-    workspace.inputs(filename) = code2
+    workspace.inputs(filename) = (code2, dialect)
     (code2, offset)
   }
