+++ tests/mtest/src/main/scala/tests/TestingWorkspaceSearch.scala
@@ -4,6 +4,7 @@ import java.nio.file.Paths
 
 import scala.collection.mutable
 
+import scala.meta.Dialect
 import scala.meta.inputs.Input
 import scala.meta.internal.metals.SemanticdbDefinition
 import scala.meta.internal.metals.WorkspaceSymbolQuery
@@ -14,21 +15,23 @@ object TestingWorkspaceSearch {
 }
 
 class TestingWorkspaceSearch {
-  val inputs: mutable.Map[String, String] = mutable.Map.empty[String, String]
+  val inputs: mutable.Map[String, (String, Dialect)] =
+    mutable.Map.empty[String, (String, Dialect)]
   def search(query: WorkspaceSymbolQuery, visitor: SymbolSearchVisitor): Unit =
     for {
-      (path, text) <- inputs
+      (path, (text, dialect)) <- inputs
     } {
-      SemanticdbDefinition.foreach(Input.VirtualFile(path, text)) { defn =>
-        if (query.matches(defn.info)) {
-          val c = defn.toCached
-          visitor.visitWorkspaceSymbol(
-            Paths.get(path),
-            c.symbol,
-            c.kind,
-            c.range
-          )
-        }
+      SemanticdbDefinition.foreach(Input.VirtualFile(path, text), dialect) {
+        defn =>
+          if (query.matches(defn.info)) {
+            val c = defn.toCached
+            visitor.visitWorkspaceSymbol(
+              Paths.get(path),
+              c.symbol,
+              c.kind,
+              c.range
+            )
+          }
       }
     }
 }
