+++ tests/unit/src/main/scala/tests/BaseWorkspaceSymbolSuite.scala
@@ -1,5 +1,7 @@
 package tests
 
+import scala.meta.Dialect
+import scala.meta.dialects
 import scala.meta.internal.metals.StatisticsConfig
 import scala.meta.internal.metals.WorkspaceSymbolProvider
 import scala.meta.io.AbsolutePath
@@ -11,9 +13,10 @@ abstract class BaseWorkspaceSymbolSuite extends BaseSuite {
   def workspace: AbsolutePath
   def statistics: StatisticsConfig = StatisticsConfig.default
   def libraries: List[Library] = Nil
+  def dialect: Dialect = dialects.Scala213
   lazy val symbols: WorkspaceSymbolProvider = {
     val p = TestingWorkspaceSymbolProvider(workspace, statistics = statistics)
-    p.indexWorkspace()
+    p.indexWorkspace(dialect)
     p.indexLibraries(libraries)
     p.indexClasspath()
     p
